requirements.txt
flask
openai
python-dotenv
zappa
